chatbot-telegram/requirements.txt
pydub
SpeechRecognition
requests
python-telegram-bot
flask
aiohttp
gunicorn
Flask[async]
python-dotenv
pytesseract
pillow
opencv-python
nest_asyncio



